drop proxy Client, merge it into Caller
