+++ app/proxy/client_test.go
@@ -12,43 +12,6 @@ import (
 	"github.com/ybbus/jsonrpc"
 )
 
-type MockRPCClient struct {
-	Delay        time.Duration
-	LastRequest  jsonrpc.RPCRequest
-	NextResponse chan *jsonrpc.RPCResponse
-}
-
-func NewMockRPCClient() *MockRPCClient {
-	return &MockRPCClient{
-		NextResponse: make(chan *jsonrpc.RPCResponse, 100),
-	}
-}
-
-func (c MockRPCClient) AddNextResponse(r *jsonrpc.RPCResponse) {
-	c.NextResponse <- r
-}
-
-func (c MockRPCClient) Call(method string, params ...interface{}) (*jsonrpc.RPCResponse, error) {
-	return <-c.NextResponse, nil
-}
-
-func (c *MockRPCClient) CallRaw(request *jsonrpc.RPCRequest) (*jsonrpc.RPCResponse, error) {
-	c.LastRequest = *request
-	return <-c.NextResponse, nil
-}
-
-func (c MockRPCClient) CallFor(out interface{}, method string, params ...interface{}) error {
-	return nil
-}
-
-func (c MockRPCClient) CallBatch(requests jsonrpc.RPCRequests) (jsonrpc.RPCResponses, error) {
-	return nil, nil
-}
-
-func (c MockRPCClient) CallBatchRaw(requests jsonrpc.RPCRequests) (jsonrpc.RPCResponses, error) {
-	return nil, nil
-}
-
 func TestClientCallDoesReloadWallet(t *testing.T) {
 	rand.Seed(time.Now().UnixNano())
 	dummyUserID := rand.Intn(100)
@@ -59,15 +22,17 @@ func TestClientCallDoesReloadWallet(t *testing.T) {
 	err = rt.UnloadWallet(dummyUserID)
 	require.NoError(t, err)
 
-	c := NewClient(rt.GetServer(dummyUserID).Address, walletID, 1*time.Second)
-
-	q, _ := NewQuery(newRawRequest(t, "wallet_balance", nil))
+	q, err := NewQuery(newRawRequest(t, "wallet_balance", nil))
+	require.NoError(t, err)
 	q.SetWalletID(walletID)
-	r, err := c.Call(q)
 
+	c := NewCaller(rt.GetServer(dummyUserID).Address, walletID)
+	r, err := c.callQueryWithRetry(q)
 	// err = json.Unmarshal(result, response)
 	require.NoError(t, err)
 	require.Nil(t, r.Error)
+
+	// TODO: check that wallet is actually reloaded? what is this test even testing?
 }
 
 func TestClientCallDoesNotReloadWalletAfterOtherErrors(t *testing.T) {
@@ -77,7 +42,7 @@ func TestClientCallDoesNotReloadWalletAfterOtherErrors(t *testing.T) {
 	srv := test.MockHTTPServer(nil)
 	defer srv.Close()
 
-	c := NewClient(srv.URL, "", 0)
+	c := NewCaller(srv.URL, "")
 	q, err := NewQuery(newRawRequest(t, "wallet_balance", nil))
 	require.NoError(t, err)
 	q.SetWalletID(walletID)
@@ -96,10 +61,9 @@ func TestClientCallDoesNotReloadWalletAfterOtherErrors(t *testing.T) {
 				Message: "Wallet at path // was not found",
 			},
 		})
-		srv.NoMoreResponses()
 	}()
 
-	r, err := c.Call(q)
+	r, err := c.callQueryWithRetry(q)
 	require.NoError(t, err)
 	require.Equal(t, "Wallet at path // was not found", r.Error.Message)
 }
@@ -108,29 +72,35 @@ func TestClientCallDoesNotReloadWalletIfAlreadyLoaded(t *testing.T) {
 	rand.Seed(time.Now().UnixNano())
 	walletID := sdkrouter.WalletID(rand.Intn(100))
 
-	mc := NewMockRPCClient()
-	c := &Client{rpcClient: mc}
-	q, _ := NewQuery(newRawRequest(t, "wallet_balance", nil))
+	srv := test.MockHTTPServer(nil)
+	defer srv.Close()
+
+	c := NewCaller(srv.URL, "")
+	q, err := NewQuery(newRawRequest(t, "wallet_balance", nil))
+	require.NoError(t, err)
 	q.SetWalletID(walletID)
 
-	mc.AddNextResponse(&jsonrpc.RPCResponse{
-		JSONRPC: "2.0",
-		Error: &jsonrpc.RPCError{
-			Message: "Couldn't find wallet: //",
-		},
-	})
-	mc.AddNextResponse(&jsonrpc.RPCResponse{
-		JSONRPC: "2.0",
-		Error: &jsonrpc.RPCError{
-			Message: "Wallet at path // is already loaded",
-		},
-	})
-	mc.AddNextResponse(&jsonrpc.RPCResponse{
-		JSONRPC: "2.0",
-		Result:  `"99999.00"`,
-	})
+	go func() {
+		srv.NextResponse <- test.ResToStr(t, jsonrpc.RPCResponse{
+			JSONRPC: "2.0",
+			Error: &jsonrpc.RPCError{
+				Message: "Couldn't find wallet: //",
+			},
+		})
+		srv.NextResponse <- "" // for the wallet_add call
+		srv.NextResponse <- test.ResToStr(t, jsonrpc.RPCResponse{
+			JSONRPC: "2.0",
+			Error: &jsonrpc.RPCError{
+				Message: "Wallet at path // is already loaded",
+			},
+		})
+		srv.NextResponse <- test.ResToStr(t, jsonrpc.RPCResponse{
+			JSONRPC: "2.0",
+			Result:  `"99999.00"`,
+		})
+	}()
 
-	r, err := c.Call(q)
+	r, err := c.callQueryWithRetry(q)
 
 	require.NoError(t, err)
 	require.Nil(t, r.Error)
